Patch undefined list retrieval order from PostgreSql
When we switched from Sqlite to PostgreSql the assumption that the database retrieves list elements in the same order as they are stored broke. This quick fix sorts the `worlds` property of the `sessionData` based on `worldId`.

Is there a database configuration setting for this, and/or should it be handled server side?

Also updates the state in the Multiworld component with new clients structure in a data immutable way, following ReactJs usage guidelines.