Guarantee at least one enemy with each prize pack
Also duplicate the prize pack for Popo2 so that it is the same as for Popo. Vanilla only uses Popo2, but just in case something like enemizer from Z3r mixes them in. This way it's future proof.

Also, all that Enumerable this and Enumerable that got to me. `using static` to the rescue.