Add "spriteinventory" and "spritegallery" commands to the CLI
The "spriteinventory" command parses the meta data block of custom sprites at a directory to produce an json data file (inventory.json) containing title and relative path for the sprites of each game.

The "spritemontage" command generates an image for each game, guided by the json data file, to be used as css sprite images for front end.