Convert from react classes to functions
React has introduced a Hooks API which we want to start using. Why? Because it simplifies using React in several ways and makes it easier to read our component code.

- Wrapper Hell
    Reusing logic involves mechanics that create big hierarchies of wrapped react components. These mechanics include:
    - Higher-order components
    - Render props
- Giant Components
    - Everything pertaining to a component needs to be encompassed by a class which makes them grow in an uncontrolled way.
    - Lifecycle methods divide functionality by time, not by tasks
- Confusing Classes
    - Even with the recently added language syntax constructs, the mechanic native to JS is that of prototype objects which makes the class concept a mismatch with the particulars of handling state and rendering of the DOM UI.
    - One example of this is the need to either use `bind`, or use the class attribute stage proposal (`prop = () => exp;` in a class), with methods that are called as event handlers.

With Hooks we can use more of a functional approach where the function is a first class citizen even toward the react framework. Do consult the following resources, in addition to the Hooks documentation itself:

(introductory) Why React Hooks? https://youtu.be/eX_L39UvZes
(in depth) React Today and Tomorrow and 90% Cleaner React With Hooks https://youtu.be/dpw9EHDh2bM