Turn NavMenu css into global styled component
Since we already have styled-components available to us we can skip importing css files in favor of grouping the css style with the component it affects. The general html selectors in NavMenu.css goes with the Layout, since it affects the page as a whole, while the other selectors relate to the NavMenu component.

Note that style-components can nest selectors just like in css processors.