Skip parseInt calls
The need to parse the input parameters came from previous functionality that allowed for manually sending items through a selection on the page. The numbers were stored as strings in the DOM and had to be parsed back to numbers.

However, this conversation should have been made "at the edge" of the application, that is in the click handler so that inner modules could "trust their neighbor", so we're skipping the `parseInt` calls in the network module. Unit and/or system tests can help catch this kind of mistake.