Persist sprite setting choices
We're JSON encoding an object under the localStorage key "persist" so that we can later persist other options.

The in order to set the index for the DropdownSelect component we had to change it to be controlled by it's parent component. If we ever need a DropdownSelect with an initial index again we can reuse the logic by calling the component function from a wrapping function that implements that behavior.