Verify files through hash comparisons
We are using the xxhash32 algorithm. Why are we using this less known algorithm?

JavaScript does not have any algorithms, secure hash or checksum, in its standard lib so we are free to choose. Here we're only after a checksum and we should value a quick procedure which still produce a high bit distribution. The xxhash algorithms have this.. in compiled code. However, every algorithm will have the disadvantage in JS of interpreted execution and the issue of floating point based numbers, so let's go with xxhash.