Improve and generalize the item bias reordering
The algorithm for reordering items based on a weight bias has the issue of not necessarily guaranteeing items are only placed beyond a certain index since the placement of later types can push prior items back beyond the weight threshold.

With this commit the algorithm is changed to consider all item types to be biased at the same time, and all types and their corresponding weights can be passed through a single method invocation.