Use the .sh scripts with Docker for smz3/sm ips builds
As well as introducing and updating the submodules for SM and combo asm, respectively.

The SM base rom is expanded to 4 mb to account for the base patch with the custom sprite implementation.

Also removed the `branch=multiworld` from the combo asm submodule entry since we should be on the master default branch now.