Resolve the traits of the supported games early
Patch had code to resolve the specific traits for the game being served by the site. This information is needed at other places in the implementation so we lift this out to the App component. From there it is passed through a react context since it's used so ubiquitously.

Since the Configure, Permalink, and Multiworld components should only service a certain game while on the corresponding sub domain we add a check for this. If there is a mismatch we offer a link with the same path, but correct hostname. The supporting functions are hosted in a new ./site module.

Together with the game traits we simplify the snes addressing such that we can skip the booleans and only use the address mapping names (exhirom and lorom).

Minor changes:
- The default game is smz3, so the trailing branch (else clause, last value in trinary operator chain) should be smz3.
- The `gameServiceUri` function (renamed `gameServiceHost`) fits with the new ./site module so we move it there.
- A `MessageCard` component is abstracted to make it easier to fit into code with multiple branches.
- In the case that `JSON.parse` throw SyntaxError, we catch and return null.