Switch to a configure route without randomizer id parameter
With the use of different domains to identify the game id, we switch over to using a general /configure path. Note that the url adjustment in Configure remains, such that the provided link will have a changed host name, but also the game id in its path. This is ok since we only deprecate the full configure path for now, waiting with removing it completely at the next major release (v12).