Update WreckedShip.cs
Previous change only considered getting through the doors of WS, not collecting items once inside, and thus put WS items in logic without the ability to get through the bomb passages to spooky missiles or Phantoon. This is a non-issue with assured early bombs, and a fringe issue in any case, but still needed fixing.