Move the Lava room location to UN West
Lava room being placed in the East region was alright previously as Super is front filled. With keysanity however, it is incorrectly placed behind a key card door.

We move the location to West instead, and duplicate the requirement for Pre-Cathedral room, as well as add logic for reaching it from Bouble Mountain. Non-HiJump walljumps on the right wall gets you back up to the door to Business Center, so the exit strategy to the left is trivial.