Clean player names
Only keep Alphanumeric and normalized (at most one consecutive) space characters. We clean the names in `Randomizer`, so that `Patch` can trust the input and focus on the conversion to snes data.

While doing this we make sure the smaller padding is on the left side when centering, to be consistent between the player names and custom sprite author credits.

Also, we add a mapping for " in `Dialog`, because why not?