Updating Normal Left Sand Pit Room requirements
Power bombs cannot get you to the top of the Left Sand Pit room. So these checks either requires morph ball, spring, or VERY fast midair morphballing. The morphballing is clearly intended for Hard Mode so this change should remove the need for this on Normal. Context: I recently had a normal run where my Spring and Bombs were Flipper locked, which turned out to be in Lower Norfair which is Space Jump, and Space Jump was in the Left Sand pit room.