Abstract rewards and medallion access into a `WorldState`
By encapsulating a list of dungeon rewards and MM/TR medallion access into a separate structure, it can be stored by the database and restored by the multiworld controller.

This needs to be separate from the location data since SM only does make use of the data. It can however make use of this in the future to store its own world state for game specific features.

`WorldState` has its own `Medallion` enum, because due to the use of `Progression` the `CanEnter` logic doesn't compare two `ItemType`s anyway. This way we differentiate the medallion items from the medallion access requirement.

Since the rewards are stored as a sequence the order of the `Region`s will be important. We take the opportunity to reorder some of them and establish "Kraid, Phantoon, Draygon, Ridley" as the canonical order between the SM bosses.

Also, we add a slight optimization in that the json serializing settings object is provided as a static field, created only once.