Update react-router-dom to its current major version
The major version 5 of React Router introduce a React hooks based interface, primarily through the `useHistory`, and `useParams` functions, which we adopt to.

Together with this change we adjust the URL params naming, pascalCase and using the term "slug" where this encoding is used to improve implicit documentation through code.

Also, removed react-router-browser as it doesn't seem to be used in the codebase.