Solve the IP key logic filler/playthrough mismatch
When the world is filled backwards we want to guarantee that the key logic is
sound. When generating a playthrough we are searching forward and following key
logic that "backtracks" could leave progression at inaccessible locations.

However, we know the filler succeeded, therefore keys can be "wasted" without
locking the intended route. We implement this by setting `World.ForwardSearch`
as true so that `CanNotWasteKeysBeforeAccessible` can adapt.

We're also restoring the logic unit test cases.