Only use data from `/api/spoiler` in `Spoiler`
Having a separate API endpoint for spoiler data implies that this is what the
`Spoiler` component should use. In reality it had been mixing using data from
the seed object passed as a prop and the data fetched from the api.

It only need to know the seed guid so this is passed down instead of the whole
seed object. Deciding to show it or not, based on the race seed setting, is
handled by the parent component instead.