Place SM rewards with bias against pendants
The rewards for SM are generated from a kind of distribution table which in
effect yield the following probabilities for the different number of pendants:
0: 66.667%
1: 30.224%
2:  2.994%
3:  0.062%

The rewards (and medallion requirements) are still stored as a sequence. I
argue that it's better to switch to storing the data as a mapping, against some
abbreviations as keys, on the next major version jump.